Style node chips
